backend/ml/requirements.txt
boto3==1.35.92
insightface==0.7.3
numpy==1.26.4
onnx==1.16.2
onnxruntime==1.16.3
opencv-python-headless==4.10.0.84
Pillow==10.4.0
